pyproject.toml
[tool.poetry]
name = "ordo-backend"
version = "0.1.0"
description = ""
authors = ["Ordo Development Team"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.6"
djangorestframework = "^3.12.4"
djangorestframework-jwt = "^1.11.0"
python-dotenv = "^0.19.0"
django-rest-passwordreset = "^1.2.0"
celery = {extras = ["sqs"], version = "^5.1.2"}
redis = "^3.5.3"
aiohttp = "^3.7.4"
Scrapy = "^2.5.0"
drf-nested-routers = "^0.93.3"
pandas = "^1.3.2"
python-slugify = "^5.0.2"
gunicorn = "^20.1.0"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
psycopg2-binary = "^2.9.1"
django-ses = "^2.2.1"
django-cors-headers = "^3.8.0"
django-storages = "^1.11.1"
Pillow = "^8.3.1"
django-filter = "^2.4.0"
factory-boy = "^3.2.0"
requests = "^2.26.0"
python-dateutil = "^2.8.2"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^5.2.0"}
django-credit-cards = "^0.4.1"
django-monthfield = "^0.1.3"
django-celery-beat = "^2.2.1"
certifi = "^2021.5.30"
django-nested-admin = "^3.3.3"
django-extensions = "^3.1.3"
aiofiles = "^0.7.0"
djangorestframework-recursive = "^0.1.2"
sentry-sdk = "^1.5.0"
stripe = "^2.63.0"
aioboto3 = "^9.3.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.14.0"
flake9 = "^3.8.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 119
extend-ignore = "E203,"
exclude = "*migrations*"

[tool.black]
line-length = 119
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
